+++ resource-registry/src/Control/ResourceRegistry.hs
@@ -344,6 +344,23 @@ nextYoungerAge (Age n) = Age (n - 1)
 data RegistryState m = RegistryState
   { registryThreads :: !(KnownThreads m)
   -- ^ Forked threads
+  , registryReleaseThreads :: ![ReleaseThread m]
+  -- ^ The list of releasing actions for threads that were forked from this
+  -- registry.
+  --
+  -- We will cancel these when closing a registry but before actually closing
+  -- the registry. This will guard the case in which a forked thread allocates
+  -- resources in the registry. We would face a race condition among:
+  --
+  -- - The registry killing the thread
+  --
+  -- - The thread allocating something in the now closed registry.
+  --
+  -- The latter case will throw an exception.
+  --
+  -- Note this is separate from 'registryThreads' because one can add threads to
+  -- cancel via 'allocateThread' and still those threads should not be directly
+  -- added to the set of known threads.
   , registryResources :: !(Map ResourceId (Resource m))
   -- ^ Currently allocated resources
   --
@@ -434,6 +451,10 @@ releaseResource Resource{resourceRelease = Release f} = f
 
 instance Show (Release m) where
   show _ = "<<release>>"
+
+-- | Release a thread when closing a registry.
+newtype ReleaseThread m = ReleaseThread {releaseThread :: m ()}
+  deriving NoThunks via OnlyCheckWhnfNamed "ReleaseThread" (ReleaseThread m)
 
 {-------------------------------------------------------------------------------
   Internal: pure functions on the registry state
@@ -598,6 +619,7 @@ initState :: RegistryState m
 initState =
   RegistryState
     { registryThreads = KnownThreads Set.empty
+    , registryReleaseThreads = []
     , registryResources = Map.empty
     , registryNextKey = ResourceId 1
     , registryAges = Bimap.empty
@@ -636,6 +658,9 @@ closeRegistry rr = mask_ $ do
         { resourceRegistryCreatedIn = registryContext rr
         , resourceRegistryUsedIn = context
         }
+
+  ts <- updateState rr $ gets registryReleaseThreads
+  mapM_ releaseThread ts
 
   -- Close the registry so that we cannot allocate any further resources
   alreadyClosed <- updateState rr $ close (contextCallStack context)
@@ -1239,6 +1264,16 @@ waitThread = wait . threadAsync
 waitAnyThread :: forall m a. MonadAsync m => [Thread m a] -> m a
 waitAnyThread ts = snd <$> waitAny (map threadAsync ts)
 
+allocateThread ::
+  (MonadMask m, MonadAsync m, HasCallStack) =>
+  ResourceRegistry m -> (ResourceId -> m (Thread m a)) -> m (ResourceKey m, Thread m a)
+allocateThread rr alloc = do
+  (k, t) <- allocate rr alloc cancelThread
+  updateState rr $
+    modify
+      (\s -> s{registryReleaseThreads = ReleaseThread (void (release k)) : registryReleaseThreads s})
+  pure (k, t)
+
 -- | Fork a new thread
 forkThread ::
   forall m a.
@@ -1250,7 +1285,7 @@ forkThread ::
   m (Thread m a)
 forkThread rr label body =
   snd
-    <$> allocate rr (\key -> mkThread key <$> async (body' key)) cancelThread
+    <$> allocateThread rr (\key -> mkThread key <$> async (body' key))
  where
   mkThread :: ResourceId -> Async m a -> Thread m a
   mkThread rid child =